Investigated transition to tracking more thoroughly and corrected a bug related to early vs. prompt code phase being required for tracking channel init.
As the Swift NAP expects the EARLY code phase to be written but the tracking channel is initialised on PROMPT rollover, we must write an early code phase of 0.5 into the channel init register to correspond to a prompt code phase of zero. Added some debug printing to look at snr and signal power druing the first 100 tracking updates. When we write the prompt code phase we see that the dll_disc is mainly positive and the snr and power are quite low indicating the dll is working hard to pull the code phase in. However when we write the early code phase (as we now expect we have to) then the dll_disc is smaller and varying closer to zero and the snr and power is much higher. This is a good indication that the tracking channel initialisation is in fact being done at the correct code phase.