Big refactor of the usart dma code. Separated into two files and I think I found a bug in the TX code.
The bug was related to getting a DMA interrupt between checking if its ok to call schedule from inside the write function and actually calling it. This caused you to access DMA regiaters when the EN bit was high (because a new transfer was initiated by the IRQ). This causes a DMA error interrupt to occur.

Not seen the freezing since I tried this but who knows if its really fixed :(