Major refactor of acquisition management.
Now stores an almanac with a debug msg callback which can receive an almanac from the console. If a valid almanac and a rough time/position estimate is available then satellites in view are calculated and searched first. If a good time estimate is available then warm starting is performed.