SOLUTION: Move process_matched_obs() back inside chMtxUnlock
Moving process_matched_obs out of the lock zone improves system liveness considerably and fixes an issue where the solution thread will miss its timing deadline. However, it would introduce an (unlikely) possibility that the low latency baseline could be calculated when the filter is in the process of being updated.

See issue #141.