TRACK: Fix scanf usage in parse_loop_params for newlib limitation
Newlib, as built for the openembedded toolchain, doesn't support the
hh size modifier in scanf (and printf? though it's unimportant there).

Thus if you use it, the byte following the location of the char you
wanted to write to will be clobbered.  Fortunately, in this case that
wasn't a problem since n_chars_read gets written after coherent_ms.