Add PPS output to Piksi's debug connector
Piksi generates a pulse-per-second synchronized with GPS time.

A thread periodically checks the current GPS time and time quality and
calculates the number of NAP clock cycles to the next full GPS second.
It then writes this value to an SPI register. Further, the pulse width
will be updated whenever the corresponding value is changed in Piksi
Console.

The PPS output has been tested in the lab with an oscilloscope using a
GPSDO as reference.